# Backend/services/dashboard-api/api/openapi.yaml
openapi: "3.0.3"
info:
  title: Zero-Trust 5G Core — Dashboard API
  description: >
    BFF for the Angular control-plane console. Contract derived from
    workflowbackend.md (card-by-card frontend analysis). All /api/v1 routes
    require a Casdoor OIDC bearer token; mutating routes additionally require
    the RBAC role noted per operation.
  version: 0.9.0
servers:
  - url: /
    description: dashboard-api (default :8093)
security:
  - bearerAuth: []
...
paths:
...
  /auth/session:
    get:
      tags: [auth]
      operationId: getSession
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Session" }
          description: Current session.
        "401": { $ref: "#/components/responses/Error" }
...
  /auth/recover:
    post:
      tags: [auth]
      security: []
      operationId: recover
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [identity, pgpSignature]
              properties:
                identity: { type: string }
                pgpSignature: { type: string, description: Detached PGP signature }
      responses:
        "202":
          content:
            application/json:
              schema:
                type: object
                properties: { recoveryToken: { type: string } }
          description: Signature verified, recovery token issued.
        "400": { $ref: "#/components/responses/Error" }

  # ───────────────────────── shell ─────────────────────────
...
  /api/v1/me/environment:
    put:
      tags: [shell]
      operationId: switchEnvironment
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [environment]
              properties: { environment: { type: string } }
      responses: { "204": { description: Environment switched. } }
...
  /api/v1/notifications:
    get:
      tags: [shell]
      operationId: listNotifications
      parameters: [{ $ref: "#/components/parameters/limit" }]
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/Notice" }
                  unread: { type: integer }
          description: Notification feed.
  /api/v1/notifications/read:
    post:
      tags: [shell]
      operationId: markNotificationsRead
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                ids: { type: array, items: { type: string }, description: Empty = all }
      responses: { "204": { description: Marked read. } }

  # ───────────────────────── dashboard / overview ─────────────────────────
...
  /api/v1/events:
    get:
      tags: [dashboard]
      operationId: listEvents
      parameters: [{ $ref: "#/components/parameters/limit" }]
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/SystemEvent" }
          description: Recent system events.
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  parameters:
...
    limit: { name: limit, in: query, schema: { type: integer, default: 20 } }
  responses:
    Error:
      description: Error envelope.
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: object
                properties:
                  code: { type: string, example: DUPLICATE_SUPI }
                  message: { type: string }
  schemas:
...
    Session:
      type: object
      properties:
        identity: { type: string, example: admin-secops@telecom.node }
        environment: { type: string, example: 5G-LAB-PROD-SOUTH }
        mtls: { type: boolean }
        since: { type: integer, format: int64 }
        role: { type: string, example: secops-admin }
        displayName: { type: string }
...
    Notice:
      type: object
      properties:
        id: { type: string }
        kind: { type: string, enum: [danger, warning, info] }
        title: { type: string }
        time: { type: string, example: 2m }
...
    SystemEvent:
      type: object
      properties:
        time: { type: string }
        severity: { type: string, enum: [critical, warning, info] }
        source: { type: string }
        message: { type: string }
...